General fixes

+++ app/middleware/attachments.js
@@ -3,9 +3,11 @@ var fs = require("fs");
 function remove(req, res, next){
 	res.on('finish', function(){
     	for(var fieldname in req.files) {
-	    	var file = req.files[fieldname]
-	    	if (!file.pending) {
-				fs.unlink(file.path);
+			if (req.files.hasOwnProperty(fieldname)) {
+				var file = req.files[fieldname]
+		    	if (!file.pending) {
+					fs.unlink(file.path);
+				}
 			}
 		}
   	});
